Se limpia auth.js y se crea controllers y middlewares
