+++ server/controllers/authController.js
@@ -1,0 +1,61 @@
+const User = require('../models/User');
+const jwt = require('jsonwebtoken');
+
+const register = async (req, res) => {
+  try {
+    const { nombre, email, password, rol, carrera } = req.body;
+
+    const usuarioExistente = await User.findOne({ email });
+    if (usuarioExistente) {
+      return res.status(409).json({ mensaje: 'Este email ya está registrado' });
+    }
+
+    const nuevoUsuario = new User({ nombre, email, password, rol, carrera });
+    await nuevoUsuario.save();
+
+    const token = jwt.sign({ id: nuevoUsuario._id, rol: nuevoUsuario.rol }, process.env.JWT_SECRET, {
+      expiresIn: '1d',
+    });
+
+    res.status(201).json({ mensaje: 'Usuario creado', token });
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({ mensaje: 'Error en el servidor' });
+  }
+};
+
+const login = async (req, res) => {
+  try {
+    const { email, password } = req.body;
+    const usuario = await User.findOne({ email });
+    if (!usuario) {
+      return res.status(401).json({ mensaje: 'Credenciales inválidas' });
+    }
+
+    const passwordValido = await usuario.compararPassword(password);
+    if (!passwordValido) {
+      return res.status(401).json({ mensaje: 'Credenciales inválidas' });
+    }
+
+    const token = jwt.sign({ id: usuario._id, rol: usuario.rol }, process.env.JWT_SECRET, {
+      expiresIn: '1d',
+    });
+
+    res.status(200).json({
+      mensaje: 'Login exitoso',
+      token,
+      usuario: {
+        id: usuario._id,
+        nombre: usuario.nombre,
+        email: usuario.email,
+        rol: usuario.rol
+      }
+    });
+
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({ mensaje: 'Error en el servidor' });
+  }
+};
+
+module.exports = { register, login };
