+++ server/index.js
@@ -1,12 +1,15 @@
 const express = require('express');
 const mongoose = require('mongoose');
 const cors = require('cors');
-const authRoutes = require('./routes/auth');
 require('dotenv').config();
 
 const app = express();
 app.use(cors());
 app.use(express.json());
+
+// Importar rutas
+
+const authRoutes = require('./routes/auth');
 app.use('/api', authRoutes);
 
 app.get('/', (req, res) => {
